Added confirm windows for when the user is about to delete a site or a program

+++ src/Components/Programs.js
@@ -57,8 +57,11 @@ var Programs = React.createClass({
 
   //deletes the selected Program
   deleteSelectedProgram() {
-    Api.deleteProgram(this.state.selectedRowId).then(() => {window.location.reload()});
-
+    var confirmed = confirm('Are you sure you want to delete this site?');
+    if (confirmed) {
+      Api.deleteProgram(this.state.selectedRowId).then(() => {window.location.reload()});
+    }
+    return;
   },
 
   render: function() {
